// src/modules/flow-xo/interfaces/flow-xo.type.ts
export interface FlowXoFlowRunDataType {
  response_path: string;
  chat_id: string;
  bot_id: string;
  bot_connection_id: string;
  data?: Record<string, any>; // user's variables
}
export interface FlowXoDefaultResponseType {
  msg: 'success';
}

export interface FlowXoRouteType {
  botId: string;
  chatId: string;
}